JS-6951: Subscription manager
