+++ src/ts/component/menu/smile.tsx
@@ -504,7 +504,7 @@ const MenuSmile = observer(class MenuSmile extends React.Component<I.Menu, State
 
 				this.setLoading(true);
 
-				U.Data.search({
+				U.Subscription.search({
 					filters,
 					sorts,
 					fullText: filter,
