+++ src/ts/component/widget/tree/index.tsx
@@ -42,7 +42,7 @@ const WidgetTree = observer(forwardRef<WidgetTreeRefProps, I.WidgetComponent>((p
 		const subIds = getSubIds();
 
 		if (subIds.length) {
-			C.ObjectSearchUnsubscribe(subIds);
+			U.Subscription.destroyList(subIds);
 			clear();
 		};
 	};
@@ -179,7 +179,7 @@ const WidgetTree = observer(forwardRef<WidgetTreeRefProps, I.WidgetComponent>((p
 		};
 
 		subscriptionHashes.current[nodeId] = hash;
-		U.Data.subscribeIds({
+		U.Subscription.subscribeIds({
 			subId,
 			ids: links,
 			keys: J.Relation.sidebar,
