+++ src/ts/lib/action.ts
@@ -50,8 +50,7 @@ class Action {
 		S.Record.metaClear(rootId, '');
 		S.Record.recordsClear(rootId, '');
 		S.Detail.clear(rootId);
-
-		C.ObjectSearchUnsubscribe([ rootId ]);
+		U.Subscription.destroyList([ rootId ]);
 	};
 
 	dbClearBlock (rootId: string, blockId: string) {
@@ -73,12 +72,10 @@ class Action {
 			S.Record.recordsClear(S.Record.getGroupSubId(rootId, blockId, id), '');
 		});
 
-		C.ObjectSearchUnsubscribe(groupIds);
-
 		S.Record.groupsClear(rootId, blockId);
 		S.Detail.clear(subId);
 
-		C.ObjectSearchUnsubscribe([ subId ]);
+		U.Subscription.destroyList(groupIds.concat([ subId ]), true);
 	};
 
 	dbClearChat (chatId: string, blockId: string) {
